lab_13_mcp_py/requirements.txt
# Lab 13 MCP server (Python): calculate, test, fetch_page_html, get_page_inputs
mcp[cli]>=1.26.0
python-dotenv>=1.0.0
requests>=2.31.0
selenium>=4.27.0
